fix destroy call to avoid NPE when group is not set. use default cert name.

+++ src/main/java/cloudos/cslib/compute/jclouds/JcloudBase.java
@@ -138,7 +138,9 @@ public class JcloudBase extends CsCloudBase {
         for (ComputeMetadata metadata : nodes) {
             if (metadata instanceof NodeMetadata) {
                 final NodeMetadata node = (NodeMetadata) metadata;
-                if (node.getGroup().startsWith(config.getGroupPrefix()) && node.getStatus() != NodeMetadata.Status.TERMINATED) {
+                if (node.getGroup() != null
+                        && node.getGroup().startsWith(config.getGroupPrefix())
+                        && node.getStatus() != NodeMetadata.Status.TERMINATED) {
                     try {
                         compute.destroyNode(node.getId());
                         destroyed++;
